+++ src/webpubsub/azext_webpubsub/_params.py
@@ -12,7 +12,8 @@ from azure.cli.core.commands.parameters import (
 )
 from .vendored_sdks.azure_mgmt_webpubsub.models import WebPubSubRequestType
 from ._actions import (
-    EventHandlerTemplateUpdateAction
+    EventHandlerTemplateUpdateAction,
+    IPRuleTemplateUpdateAction
 )
 from ._validator import validate_network_rule
 
@@ -52,6 +53,7 @@ def load_arguments(self, _):
         c.argument('public_network', arg_type=get_three_state_flag(), help='Set rules for public network.', required=False, arg_group='Public Network')
         c.argument('allow', arg_type=get_enum_type(WebPubSubRequestType), nargs='*', help='The allowed virtual network rule. Space-separeted list of scope to assign.', type=WebPubSubRequestType, required=False)
         c.argument('deny', arg_type=get_enum_type(WebPubSubRequestType), nargs='*', help='The denied virtual network rule. Space-separeted list of scope to assign.', type=WebPubSubRequestType, required=False)
+        c.argument('ip_rule', action=IPRuleTemplateUpdateAction, nargs='*', help='The IP rule for the hub.', required=False)
 
     for scope in ['webpubsub hub delete',
                   'webpubsub hub show']:
@@ -64,6 +66,7 @@ def load_arguments(self, _):
             c.argument('hub_name', help='The hub to manage')
             c.argument('event_handler', action=EventHandlerTemplateUpdateAction, nargs='*', help='Template item for event handler settings. Use key=value pattern to set properties. Supported keys are "url-template", "user-event-pattern", "system-event", "auth-type" and "auth-resource". Setting multiple "system-event" results in an array and for other properties, only last set takes active.')
             c.argument('allow_anonymous', arg_type=get_three_state_flag(), help='Set if anonymous connections are allowed for this hub. True means allow and False means deny.')
+            c.argument('websocket_keepalive', help='The WebSocket keep-alive interval in seconds for all clients in the hub. Valid range: 1 to 120. Default to 20 seconds.')
 
     with self.argument_context('webpubsub hub list') as c:
         c.argument('webpubsub_name', webpubsub_name_type, options_list=['--name', '-n'], id_part=None)
@@ -108,7 +111,10 @@ def load_arguments(self, _):
     for scope in ['webpubsub replica create',
                   'webpubsub replica list',
                   'webpubsub replica delete',
-                  'webpubsub replica show']:
+                  'webpubsub replica show',
+                  'webpubsub replica start',
+                  'webpubsub replica stop',
+                  'webpubsub replica restart']:
         with self.argument_context(scope) as c:
             c.argument('sku', help='The sku name of the replica. Currently allowed values: Premium_P1')
             c.argument('unit_count', help='The number of webpubsub service unit count', type=int)