+++ src/webpubsub/azext_webpubsub/commands.py
@@ -57,6 +57,8 @@ def load_command_table(self, _):
         g.command('delete', 'webpubsub_delete')
         g.command('list', 'webpubsub_list')
         g.show_command('show', 'webpubsub_show', exception_handler=empty_on_404)
+        g.command('start', 'webpubsub_start', exception_handler=empty_on_404)
+        g.command('stop', 'webpubsub_stop', exception_handler=empty_on_404)
         g.command('restart', 'webpubsub_restart', exception_handler=empty_on_404)
         g.generic_update_command('update', getter_name='webpubsub_get',
                                  setter_name='webpubsub_set',
@@ -110,4 +112,7 @@ def load_command_table(self, _):
         g.command('create', 'webpubsub_replica_create')
         g.command('list', 'webpubsub_replica_list')
         g.show_command('show', 'webpubsub_replica_show', exception_handler=empty_on_404)
+        g.command('start', 'webpubsub_replica_start', exception_handler=empty_on_404)
+        g.command('stop', 'webpubsub_replica_stop', exception_handler=empty_on_404)
+        g.command('restart', 'webpubsub_replica_restart', exception_handler=empty_on_404)
         g.show_command('delete', 'webpubsub_replica_delete')